requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
pymupdf>=1.24.0        # slide PDF parsing (fitz)
pydantic>=2.0.0
python-dotenv>=1.0.0
